requirements.txt
streamlit
langchain
python-dotenv
openai
